+++ code/extensions/che-api/src/impl/k8s-service-impl.ts
@@ -17,9 +17,11 @@ import {
 } from '@devfile/api';
 import * as k8s from '@kubernetes/client-node';
 
+import * as vscode from 'vscode';
 import { ApiType } from '@kubernetes/client-node';
 import { injectable } from 'inversify';
 import { K8SRawResponse, K8SService } from '../api/k8s-service';
+import { env } from 'process';
 
 const request = require('request');
 
@@ -34,6 +36,36 @@ export class K8SServiceImpl implements K8SService {
   constructor() {
     this.k8sConfig = new k8s.KubeConfig();
     this.k8sConfig.loadFromCluster();
+  }
+
+  async ensureKubernetesServiceHostWhitelisted(): Promise<void> {
+    const proxy = env.HTTPS_PROXY || env.HTTP_PROXY || env.https_proxy || env.http_proxy;
+    const noProxy = env.NO_PROXY || env.no_proxy;
+    if (proxy && noProxy && env.KUBERNETES_SERVICE_HOST) {
+
+      // take k8s service host
+      const k8sHost = env.KUBERNETES_SERVICE_HOST;
+
+      // check whether it is set to no_proxy environment variable
+      if (!noProxy.split(',').includes(k8sHost)) {
+        const action = await vscode.window.showInformationMessage(
+          'The cluster you are using is behind a proxy, but kubernetes service host is not whitelisted in no_proxy environment variable. ' +
+          'This may cause the kubernetes service to be unaccessible. ' +
+          'Do you want to fix this and add the kubernetes service host to the no_proxy environment variable?', 'Add', 'Cancel');
+
+        if ('Add' === action) {
+          if (env.NO_PROXY) {
+            env.NO_PROXY += ',' + k8sHost;
+            console.log('Kubernetes Service Host has been added to env.NO_PROXY environment variable');
+          }
+          if (env.no_proxy) {
+            env.no_proxy += ',' + k8sHost;
+            console.log('Kubernetes Service Host has been added to env.no_proxy environment variable');
+          }
+        }
+      }
+
+    }
   }
 
   // eslint-disable-next-line @typescript-eslint/no-explicit-any