+++ code/src/vs/base/common/codiconsLibrary.ts
@@ -579,4 +579,10 @@ export const codiconsLibrary = {
 	percentage: register('percentage', 0xec33),
 	sortPercentage: register('sort-percentage', 0xec33),
 	attach: register('attach', 0xec34),
+	goToEditingSession: register('go-to-editing-session', 0xec35),
+	editSession: register('edit-session', 0xec36),
+	codeReview: register('code-review', 0xec37),
+	copilotWarning: register('copilot-warning', 0xec38),
+	python: register('python', 0xec39),
+	copilotLarge: register('copilot-large', 0xec3a),
 } as const;