+++ .rebase/CHANGELOG.md
@@ -2,6 +2,29 @@
 
 The file to keep a list of changed files which will potentionaly help to resolve rebase conflicts.
 
+#### @RomanNikitenko
+https://github.com/che-incubator/che-code/pull/463
+
+- code/package.json
+- code/test/automation/package.json
+- code/test/smoke/package.json
+---
+
+#### @RomanNikitenko
+https://github.com/che-incubator/che-code/pull/456
+
+- code/src/vs/workbench/browser/parts/titlebar/commandCenterControl.ts
+- code/src/vs/workbench/browser/parts/titlebar/titlebarPart.ts
+- code/src/vs/workbench/browser/parts/titlebar/windowTitle.ts
+- code/src/vs/workbench/browser/workbench.contribution.ts
+---
+
+#### @RomanNikitenko
+https://github.com/che-incubator/che-code/pull/460
+
+- code/remote/package.json
+---
+
 #### @benoitf @RomanNikitenko
 https://github.com/che-incubator/che-code/pull/379 \
 https://github.com/che-incubator/che-code/commit/d3cf7dc86d284bc4cdff7cc163c5642bb6744524#diff-36f85da944a1b6c01cb656687e520f7415d692e1b7d8856e2161db97a134b224
@@ -78,4 +101,10 @@ https://github.com/che-incubator/che-code/pull/193
 https://github.com/che-incubator/che-code/pull/353
 
 - code/src/vs/server/node/extensionHostConnection.ts
+---
+
+#### @vinokurig
+https://github.com/che-incubator/che-code/pull/400
+
+- code/extensions/git/src/ssh-askpass.sh
 ---